+++ Infrastructure/Services/IncomeExpenseService.cs
@@ -1,6 +1,7 @@
 using AutoMapper;
 using Domain.Entities;
 using Domain.Repository;
+using Infrastructure.Models.Exceptions;
 using Infrastructure.Models.Requests;
 using Infrastructure.Models.Responses;
 
@@ -24,11 +25,26 @@ public class IncomeExpenseService : IIncomeExpenseService
 
     public async Task<IncomeExpenseCategory> GetIncomeExpenseType(int id)
     {
-        return await _incomeExpensesRepository.Get(id);
+        if (id <=0)
+        {
+            throw new BadRequestException(BadRequestException.WrongIdMessage);
+        }
+        var result = await _incomeExpensesRepository.Get(id);
+
+        if (result == null)
+        {
+            throw new NotFoundException();
+        }
+
+        return result;
     }
 
     public async Task<IncomeExpensesAddResponse> CreateIncomeExpenseType(IncomeExpensesAddRequest request)
     {
+        if (request == null)
+        {
+            throw new BadRequestException();
+        }
         IncomeExpenseCategory entityCategory = _mapper.Map<IncomeExpensesAddRequest, IncomeExpenseCategory>(request);
         
         IncomeExpensesAddResponse response = _mapper.Map<IncomeExpenseCategory , IncomeExpensesAddResponse>(await _incomeExpensesRepository.Create(entityCategory));
@@ -40,16 +56,26 @@ public class IncomeExpenseService : IIncomeExpenseService
     {
         if (id <= 0)
         {
-            throw new ArgumentOutOfRangeException(nameof(id));
+            throw new BadRequestException(BadRequestException.WrongIdMessage);
         }
 
         var type = _incomeExpensesRepository.Get(id);
+
+        if (type.Result == null)
+        {
+            throw new NotFoundException();
+        }
 
         _incomeExpensesRepository.Remove(type.Result);
     }
 
     public async Task<IncomeExpensesUpdateResponse> UpdateIncomeExpenseType(IncomeExpensesUpdateRequest request)
     {
+        if (request == null)
+        {
+            throw new BadRequestException();
+        }
+        
         IncomeExpenseCategory entityCategory = _mapper.Map<IncomeExpensesUpdateRequest, IncomeExpenseCategory>(request);
         
         IncomeExpensesUpdateResponse response = _mapper.Map<IncomeExpenseCategory , IncomeExpensesUpdateResponse>(await _incomeExpensesRepository.Update(entityCategory));
